+++ waspc/cli/src/Wasp/Cli/Command/Start/ArgumentsParser.hs
@@ -1,0 +1,18 @@
+module Wasp.Cli.Command.Start.ArgumentsParser
+  ( StartArgs (..),
+    startArgsParser,
+  )
+where
+
+import Network.Socket (PortNumber)
+import qualified Options.Applicative as Opt
+import Wasp.Cli.Util.PortArgument (servicePortsParser)
+import Wasp.Project.PerService (PerService (..))
+
+newtype StartArgs = StartArgs
+  { ports :: PerService PortNumber
+  }
+  deriving (Eq, Show)
+
+startArgsParser :: Opt.Parser StartArgs
+startArgsParser = StartArgs <$> servicePortsParser